Move lazyprop to the utils module
